Gracefully fail on large numbers of VIP changes
I recently ran into a must-gather with 10s of thousands of VIP changes
on one node. Although the code processed it fine, the UI completely
hung when trying to display that many separate items. Now when we
encounter more than 1000 (an educated guess for how many VIP changes
we would expect before we just assume the node is hosed) changes
we simply display an error label.